Mlops-stock prediction/vertex_ai/requirement.txt
streamlit 
yfinance
plotly
google-cloud
google-auth
google-auth-oauthlib
google-cloud-aiplatform
